+++ sql/parser/create_test.go
@@ -22,14 +22,14 @@ func TestParserCreateTable(t *testing.T) {
 			query.CreateTableStmt{
 				TableName: "test",
 				Config: database.TableConfig{
-					PrimaryKey: database.FieldConstraint{Path: []string{"foo"}, Type: document.IntValue},
+					PrimaryKey: database.FieldConstraint{Path: []string{"foo"}, Type: document.Int64Value},
 				},
 			}, false},
 		{"With multiple constraints", "CREATE TABLE test(foo INT PRIMARY KEY, bar INT16, baz.4.1.bat STRING)",
 			query.CreateTableStmt{
 				TableName: "test",
 				Config: database.TableConfig{
-					PrimaryKey: database.FieldConstraint{Path: []string{"foo"}, Type: document.IntValue},
+					PrimaryKey: database.FieldConstraint{Path: []string{"foo"}, Type: document.Int64Value},
 					FieldConstraints: []database.FieldConstraint{
 						{Path: []string{"bar"}, Type: document.Int16Value},
 						{Path: []string{"baz", "4", "1", "bat"}, Type: document.StringValue},
